+++ pkgs/package_config/test/package_config_impl_test.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import "dart:convert" show jsonDecode;
+
 import "package:package_config/package_config_types.dart";
 import "package:test/test.dart";
 import "src/util.dart";
@@ -72,6 +74,7 @@ void main() {
       var absolute = root.resolve("foo/bar/");
       var package = Package("name", root,
           packageUriRoot: absolute,
+          relativeRoot: false,
           languageVersion: version,
           extraData: unique);
       expect(package.name, "name");
@@ -79,16 +82,18 @@ void main() {
       expect(package.packageUriRoot, absolute);
       expect(package.languageVersion, version);
       expect(package.extraData, same(unique));
+      expect(package.relativeRoot, false);
     });
 
     test("relative package root", () {
       var relative = Uri.parse("foo/bar/");
       var absolute = root.resolveUri(relative);
-      var package =
-          Package("name", root, packageUriRoot: relative, extraData: unique);
+      var package = Package("name", root,
+          packageUriRoot: relative, relativeRoot: true, extraData: unique);
       expect(package.name, "name");
       expect(package.root, root);
       expect(package.packageUriRoot, absolute);
+      expect(package.relativeRoot, true);
       expect(package.languageVersion, null);
       expect(package.extraData, same(unique));
     });
@@ -138,6 +143,44 @@ void main() {
       expect(single.resolve(pkg("a", "b")), isNull);
       var resolved = single.resolve(pkg("name", "a/b"));
       expect(resolved, root.resolve("a/b"));
+    });
+  });
+  test("writeString", () {
+    var config = PackageConfig([
+      Package("foo", Uri.parse("file:///pkg/foo/"),
+          packageUriRoot: Uri.parse("file:///pkg/foo/lib/"),
+          relativeRoot: false,
+          languageVersion: LanguageVersion(2, 4),
+          extraData: {"foo": "foo!"}),
+      Package("bar", Uri.parse("file:///pkg/bar/"),
+          packageUriRoot: Uri.parse("file:///pkg/bar/lib/"),
+          relativeRoot: true,
+          extraData: {"bar": "bar!"}),
+    ], extraData: {
+      "extra": "data"
+    });
+    var buffer = StringBuffer();
+    PackageConfig.writeString(config, buffer, Uri.parse("file:///pkg/"));
+    var text = buffer.toString();
+    var json = jsonDecode(text); // Is valid JSON.
+    expect(json, {
+      "configVersion": 2,
+      "packages": unorderedEquals([
+        {
+          "name": "foo",
+          "rootUri": "file:///pkg/foo/",
+          "packageUri": "lib/",
+          "languageVersion": "2.4",
+          "foo": "foo!",
+        },
+        {
+          "name": "bar",
+          "rootUri": "bar/",
+          "packageUri": "lib/",
+          "bar": "bar!",
+        },
+      ]),
+      "extra": "data",
     });
   });
 }